encrypted-keys: style and other cleanup
Cleanup based on David Howells suggestions:
- use static const char arrays instead of #define
- rename init_sdesc to alloc_sdesc
- convert 'unsigned int' definitions to 'size_t'
- revert remaining 'const unsigned int' definitions to 'unsigned int'

Acked-by: David Howells <dhowells@redhat.com>

+++ security/keys/encrypted_defined.h
@@ -5,8 +5,7 @@ FASTVC-BENCH-CORPUS:linux-main-100k-v1-e0bd41dc-8e12-4f1b-978d-a3645ae59da0
 #define ENCRYPTED_DEBUG 0
 
 #if ENCRYPTED_DEBUG
-static inline void dump_master_key(const u8 *master_key,
-				   unsigned int master_keylen)
+static inline void dump_master_key(const u8 *master_key, size_t master_keylen)
 {
 	print_hex_dump(KERN_ERR, "master key: ", DUMP_PREFIX_NONE, 32, 1,
 		       master_key, master_keylen, 0);
@@ -35,8 +34,7 @@ static inline void dump_hmac(const char *str, const u8 *digest,
 		       hmac_size, 0);
 }
 #else
-static inline void dump_master_key(const u8 *master_key,
-				   unsigned int master_keylen)
+static inline void dump_master_key(const u8 *master_key, size_t master_keylen)
 {
 }
